Remove AnimatedSwitch component and refactor related imports; update Switch component to include animated functionality and theme-aware styles. Adjust SettingsToggle to utilize the new Switch implementation for improved consistency across components.
