+++ src/components/atoms/index.ts
@@ -3,7 +3,6 @@ export { IconButton } from './IconButton';
 export { Text } from './Text';
 export { TextButton, CircleTextButton } from './TextButton';
 export { Switch } from './Switch';
-export { AnimatedSwitch } from './AnimatedSwitch';
 export { LoadingSpinner } from './LoadingSpinner';
 export { Logo } from './logo';
 export { SafeAreaView } from './SafeAreaView';
