+++ src/components/organisms/Instructions/InstructionsModal.tsx
@@ -6,9 +6,10 @@ import { InstructionMode } from '@types';
 
 /**
  * Props for the InstructionsModal component.
- * @property mode - The instruction mode that determines which instructions to show
+ * @property mode - The instruction mode that determines which instructions to show and their presentation
  */
 interface InstructionsModalProps {
+  /** The instruction mode that determines which instructions to show and their presentation */
   mode: InstructionMode;
 }
 
@@ -19,11 +20,23 @@ interface InstructionsModalProps {
  * the modal visibility state. It automatically shows instructions when appropriate
  * and tracks when the user has dismissed instructions to avoid showing them again.
  *
+ * Features:
+ * - Modal dialog presentation with consistent styling
+ * - Automatic instruction dismissal tracking
+ * - Mode-based instruction content selection
+ * - Integration with page modal context for visibility management
+ * - Session state persistence for dismissed instructions
+ * - Conditional title display based on instruction mode
+ *
  * @param props - InstructionsModalProps containing the instruction mode
  * @returns JSX element containing the instructions modal
  *
  * @example
  * ```typescript
+ * // Show all instructions
+ * <InstructionsModal mode={InstructionMode.ALL} />
+ *
+ * // Show specific instruction mode
  * <InstructionsModal mode={InstructionMode.EDIT} />
  * ```
  */
